pyproject.toml
[project]
name = "ai-trading"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
app = ["fastapi","uvicorn","pydantic","sqlalchemy","python-dotenv"]
research = ["pandas","numpy","pyarrow","pyyaml","tqdm","prefect","hydra-core","scikit-learn","pytest","ruff"]

[tool.ruff]
line-length = 100
